Improvements to template:
Ability to save and restore context using "with" sections:

    {% set a = 100 %}
    {% with %}
        {% set a = 200 %}
        {{ a }}
    {% endwith %}
    {{ a }}

Ability to include another file for a template at run time. The included
file will receive a copy of the current context.  Currently there is no
checking on the filename other than ensuring it is a string.

    {% include path/to/file.tmpl %}

Line numbers are reported in syntax errors.