Minor changes to space processing.
The idea is, when dealing with space, it is reasonable to expect that a
function will skip space at the beginning, but not afterwards.  So you
should be able to call _parse_expr without calling _skip_space first,
but after parsing the expression, if you need anything special you
should skip the space.