Include now specifies return var, increased version.
The include syntax has been changed to support specifying the variable
to assign the return dict to.  If not specified, return values will
be discarded:

    {% include "template" %}
    {% include "template" with var=value, var2=value2 %}
    {% include "template" return results %}
    {% include "template" return results with var=value, var2=value2 %}

The return variable is a dictionary with the names assigned from any
return action in the called template.  So if the called template
executes:

    {% return var=1,var2="20" %}
    {% return name="Test" %}

Then the calling template can perform the following:

    {% include "template" return result %}

    {{ result.var1 }}
    {{ result.var2 }}
    {{ retult.name }}