Changed parsing behavior.
Escapes are no longer recognized in the test. Instead, during parsing,
if it sees a non-tag opening brace, it just ignores it and passes it on
through as text.  This enables creating templates for items that need
literal braces in the output simpler.  To emit a literal tag, just
include it in an emitter tag: {{ "{{" }}