Major overhaul of how variables are stored and accessed.
Variables are now classified into three types, local, global, and
private.  Variables beginning and ending with an underscore default to
global.  Variables only beginning with an underscore, and the underscore
itself, default to private. All others default to local.

Variables are only accessed within the compartment it is set as. So if a
variable is set as global, it cannot be accessed as local or private.
Local variables are copied down each time a new template is included.
Nested templates can inherit the local variables set from upper
templates, but changes to those variables will not be visible when
returning to the upper templates.  Private variables are never copied
down, so any private variables set or retrieved will not interfere with
other local, global, or private variables. and cannot be changed from
other included templates.  Local variables of a template also cannot be
changed from other included templates, but interfere by means of being
copied down to those templates.

Because of how they are accessed, if a variable is set as global, it
must be accessed as global. Explicit access can be done by prefixing a
varible name with the type code followed by "@", where the type codes
are l for local, g for global, and p for private.  This will override
the inferred type from leading/trailing underscores:

    _ = private
    _a_variable = private
    _a_variable_ = global
    a_variable = local
    a_variable_ = local
    __ = global (two underscores)

Variable types are used in set and unset tags, import tags, the return
variables for include and code, the item and index variable in an
iterative for, and the init and increment variables in an incremental
for.

    {% set x = 1, _y = 2, g@_z = 3 %}

x is local, _y is private, and _z is global. _y is not available to
included template or parent templates, though those templates may have
their own private _y variable. x is available to included template, and
those templates can even change their own value of x, but when rendering
returns to the current template, it will still have the original value
of x. _z is a global and is available to all template, but must be
accessed as a global.  It cannot just be accessed as "_z" as that
defaults to private, so any template using it must use "g@_z". If a
global is inferred via leading and trailing underscore, the explicit
form isn't needed.