Minor changes to code using parser._find_tag_segments.
For code that uses find_tag_segments, expect the expression to be the
first segment, then the other segmetns may contain other items. The
reason is "with", "return", "into", etc are all valid variable in a
template, so by forcing the first item to be the expression, they can be
used as variables:

    {% set with = "filename.tmpl" %}
    {% include with ; with var = 5 %}

This would have failed before, as the include action's handler would see
the first "with" and treat it as the start of a variable list instead,
with no way to access it as a variable.